+++ examplelatest/src/Examples/DarkEditor.tsx
@@ -36,18 +36,9 @@ export const DarkEditor = ({}: NativeStackScreenProps<any, any, any>) => {
 
   return (
     <SafeAreaView
-      style={{
-        ...exampleStyles.fullScreen,
-        backgroundColor: EDITOR_BACKGROUND_COLOR,
-      }}
+      style={[exampleStyles.fullScreen, exampleStyles.container]}
     >
-      <View
-        style={{
-          ...exampleStyles.fullScreen,
-          paddingHorizontal: 12,
-          backgroundColor: EDITOR_BACKGROUND_COLOR,
-        }}
-      >
+      <View style={[exampleStyles.fullScreen, exampleStyles.editorContent]}>
         <RichText editor={editor} />
       </View>
       <KeyboardAvoidingView
@@ -82,6 +73,13 @@ const ToolbarWithColor = ({ editor }: ToolbarWithColorProps) => {
 const exampleStyles = StyleSheet.create({
   fullScreen: {
     flex: 1,
+  },
+  container: {
+    backgroundColor: EDITOR_BACKGROUND_COLOR
+  },
+  editorContent: {
+    paddingHorizontal: 12,
+    backgroundColor: EDITOR_BACKGROUND_COLOR,
   },
   keyboardAvoidingView: {
     position: 'absolute',