+++ src/bridges/image.ts
@@ -1,10 +1,13 @@
-import Image from '@tiptap/extension-image';
+import Image, {
+  type ImageOptions,
+  type SetImageOptions,
+} from '@tiptap/extension-image';
 import BridgeExtension from './base';
 
 type ImageEditorState = {};
 
 type ImageEditorInstance = {
-  setImage: (src: string) => void;
+  setImage: (options: SetImageOptions) => void;
 };
 
 declare module '../types/EditorBridge' {
@@ -18,13 +21,14 @@ export enum ImageEditorActionType {
 
 type ImageMessage = {
   type: ImageEditorActionType.SetImage;
-  payload: string;
+  payload: SetImageOptions;
 };
 
 export const ImageBridge = new BridgeExtension<
   ImageEditorState,
   ImageEditorInstance,
-  ImageMessage
+  ImageMessage,
+  ImageOptions
 >({
   tiptapExtension: Image.configure({
     allowBase64: true,
@@ -34,7 +38,7 @@ export const ImageBridge = new BridgeExtension<
       editor
         .chain()
         .focus()
-        .setImage({ src: message.payload })
+        .setImage(message.payload)
         .setTextSelection(editor.state.selection.to + 1)
         .run();
     }
@@ -43,10 +47,10 @@ export const ImageBridge = new BridgeExtension<
   },
   extendEditorInstance: (sendBridgeMessage) => {
     return {
-      setImage: (src: string) =>
+      setImage: (options: SetImageOptions) =>
         sendBridgeMessage({
           type: ImageEditorActionType.SetImage,
-          payload: src,
+          payload: options,
         }),
     };
   },