+++ src/bridges/trailingNode.ts
@@ -1,0 +1,6 @@
+import { TrailingNode } from '@tiptap/extensions';
+import BridgeExtension from './base';
+
+export const TrailingNodeBridge = new BridgeExtension({
+  tiptapExtension: TrailingNode,
+});